+++ cg/services/analysis_starter/configurator/extensions/raredisease.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from cg.services.analysis_starter.configurator.extensions.abstract import PipelineExtension
 from cg.services.analysis_starter.configurator.file_creators.gene_panel import GenePanelFileCreator
-from cg.services.analysis_starter.configurator.file_creators.nextflow.managed_variants import (
+from cg.services.analysis_starter.configurator.file_creators.managed_variants import (
     ManagedVariantsFileCreator,
 )
