+++ tests/fixture_plugins/analysis_starter/extension_fixtures.py
@@ -2,7 +2,7 @@ import pytest
 
 from cg.services.analysis_starter.configurator.extensions.raredisease import RarediseaseExtension
 from cg.services.analysis_starter.configurator.file_creators.gene_panel import GenePanelFileCreator
-from cg.services.analysis_starter.configurator.file_creators.nextflow.managed_variants import (
+from cg.services.analysis_starter.configurator.file_creators.managed_variants import (
     ManagedVariantsFileCreator,
 )
